Created the GUI shell that needs filling with actions
The UI is now run from moxui.py it has everything I need to create the application interface. I now need to implement the actions of the generator into it. Might use the generator.py as a seperate module for generating the QR code as well as uploading the file. Need to also look into a good way to make the progress bar representative of accurate feedback.